+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Table component: added option "highlight selected row"
 - Display a comments on the waterfall chart
+- Added visualisation "risk-review"
 
 ### Changed
 - DatePicker intervals moves with date